Dev savindu

+++ src/main/java/com/inova/project_manager_api/services/ResourceService.java
@@ -5,5 +5,5 @@ import com.inova.project_manager_api.utils.StandardResponse;
 public interface ResourceService {
     StandardResponse searchPotentialResources();
 
-    StandardResponse getEmployeesAndProjectsNotAllocatedToProject(Long projectId);
+    StandardResponse getEmployeesAndProjectsNotAllocatedToProject(int projectId);
 }
